+++ Engine/src/engine.c
@@ -191,11 +191,11 @@ uint8_t  globparaceilclip, globparaflorclip;
 int32_t xyaspect, viewingrangerecip;
 
 int32_t asm1, asm4;
-intptr_t asm2, asm3;
+long asm2, asm3;
 
 
 int32_t vplce[4], vince[4];
-intptr_t bufplce[4];
+long bufplce[4];
 
 uint8_t*  palookupoffse[4];
 
@@ -1226,7 +1226,7 @@ static void wallscan(int32_t x1, int32_t x2,
                      int32_t *swal, int32_t *lwal)
 {
     int32_t x, xnice, ynice;
-    intptr_t i;
+    long i;
     uint8_t* fpalookup;
     int32_t y1ve[4], y2ve[4], u4, d4, z, tileWidth, tsizy;
     uint8_t  bad;
@@ -1416,7 +1416,7 @@ static void maskwallscan(int32_t x1, int32_t x2,
                          int32_t *swal, int32_t *lwal)
 {
     int32_t x, startx, xnice, ynice;
-    intptr_t i;
+    long i;
     uint8_t* fpalookup;
     int32_t y1ve[4], y2ve[4], u4, d4, dax, z, tileWidth, tileHeight;
     uint8_t*  p;
@@ -3080,7 +3080,7 @@ static void transmaskvline(int32_t x)
 static void transmaskvline2 (int32_t x)
 {
     int32_t y1, y2, x2;
-    intptr_t i;
+    long i;
     short y1ve[2], y2ve[2];
 
     if ((x < 0) || (x >= xdimen)) return;
@@ -8996,7 +8996,7 @@ void drawmapview(int32_t dax, int32_t day, int32_t zoome, short ang)
 void clearview(int32_t dacol)
 {
     int32_t y, dx;
-    intptr_t p;
+    long p;
 
     if (qsetmode != 200) return;
 
@@ -9109,7 +9109,7 @@ void preparemirror(int32_t dax, int32_t day, int32_t daz,
 void completemirror(void)
 {
     int32_t i, dy;
-    intptr_t p;
+    long p;
 
     /* Can't reverse with uninitialized data */
     if (inpreparemirror) {
